Work around bad diagnostic when missing Context import
Previously the compiler would say:

    error[E0599]: no method named `context` found for type `std::result::Result<(), anyhow::error::Error>` in the current scope
      --> src/main.rs:83:13
       |
    83 |             context(...)?;
       |             ^^^^^^^ method not found in `std::result::Result<(), anyhow::error::Error>`
       |
       = help: items from traits can only be used if the trait is in scope
    help: the following trait is implemented but not in scope; perhaps add a `use` for it:
       |
    19 | use anyhow::context::Context;

where the `anyhow::context::Context` in the suggestion is not a path
that is accessible to the user.
